+++ tests/unit/nn/wrapper_test.py
@@ -13,7 +13,7 @@ class TestTensorOps(unittest.TestCase):
     def test_mul_numbers(self):
         constant, shape = 10, (3, 5)
         with nn.variable_scope(self.id().replace('.', '/')):
-            tensor1 = fixture.create_tensor(shape)
+            tensor1 = fixture.create_tensor(shape, dtype='int32')
             tensor2 = tensor1 * constant
             tensor3 = constant * tensor1
 
@@ -28,7 +28,7 @@ class TestTensorOps(unittest.TestCase):
     def test_mul_tensor(self):
         shape = (3, 5)
         with nn.variable_scope(self.id().replace('.', '/')):
-            tensor1 = fixture.create_tensor(shape)
+            tensor1 = fixture.create_tensor(shape, dtype='int32')
             tensor2 = tensor1 * tensor1
 
         session = nn.Session()
@@ -41,8 +41,8 @@ class TestTensorOps(unittest.TestCase):
     def test_mul_variable(self):
         shape = (3, 5)
         with nn.variable_scope(self.id().replace('.', '/')):
-            tensor1 = fixture.create_tensor(shape)
-            variable = fixture.create_variable(shape)
+            tensor1 = fixture.create_tensor(shape, dtype='int32')
+            variable = fixture.create_variable(shape, dtype='int32')
             tensor2 = tensor1 * variable
             tensor3 = variable * tensor1
 
@@ -58,7 +58,7 @@ class TestTensorOps(unittest.TestCase):
     def test_mul_input(self):
         shape = (3, 5)
         with nn.variable_scope(self.id().replace('.', '/')):
-            tensor1 = fixture.create_tensor(shape)
+            tensor1 = fixture.create_tensor(shape, dtype='int32')
             input_ = nn.Input(name='input', shape=shape, dtype='int32')
             tensor2 = tensor1 * input_
             tensor3 = input_ * tensor1
@@ -77,7 +77,7 @@ class TestTensorOps(unittest.TestCase):
     def test_add_numbers(self):
         constant, shape = 10, (3, 5)
         with nn.variable_scope(self.id().replace('.', '/')):
-            tensor1 = fixture.create_tensor(shape)
+            tensor1 = fixture.create_tensor(shape, dtype='int32')
             tensor2 = tensor1 + constant
             tensor3 = constant + tensor1
 
@@ -92,7 +92,7 @@ class TestTensorOps(unittest.TestCase):
     def test_add_tensor(self):
         shape = (3, 5)
         with nn.variable_scope(self.id().replace('.', '/')):
-            tensor1 = fixture.create_tensor(shape)
+            tensor1 = fixture.create_tensor(shape, dtype='int32')
             tensor2 = tensor1 + tensor1
 
         session = nn.Session()
@@ -105,7 +105,7 @@ class TestTensorOps(unittest.TestCase):
         shape = (3, 5)
         with nn.variable_scope(self.id().replace('.', '/')):
             input_ = nn.Input(shape=shape, dtype='int32')
-            tensor1 = fixture.create_tensor(shape)
+            tensor1 = fixture.create_tensor(shape, dtype='int32')
             tensor2 = tensor1 + input_
             tensor3 = input_ + tensor1
 
@@ -121,8 +121,8 @@ class TestTensorOps(unittest.TestCase):
     def test_add_variable(self):
         shape = (3, 5)
         with nn.variable_scope(self.id().replace('.', '/')):
-            tensor1 = fixture.create_tensor(shape)
-            variable = fixture.create_variable(shape)
+            tensor1 = fixture.create_tensor(shape, dtype='int32')
+            variable = fixture.create_variable(shape, dtype='int32')
             tensor2 = tensor1 + variable
             tensor3 = variable + tensor1
 
@@ -138,7 +138,7 @@ class TestTensorOps(unittest.TestCase):
     def test_neg(self):
         shape = (3, 5)
         with nn.variable_scope(self.id().replace('.', '/')):
-            tensor1 = fixture.create_tensor(shape)
+            tensor1 = fixture.create_tensor(shape, dtype='int32')
             tensor2 = -tensor1
 
         session = nn.Session()
@@ -150,7 +150,7 @@ class TestTensorOps(unittest.TestCase):
     def test_sub_numbers(self):
         constant, shape = 10, (3, 5)
         with nn.variable_scope(self.id().replace('.', '/')):
-            tensor1 = fixture.create_tensor(shape)
+            tensor1 = fixture.create_tensor(shape, dtype='int32')
             tensor2 = tensor1 - constant
             tensor3 = constant - tensor1
 
@@ -165,7 +165,7 @@ class TestTensorOps(unittest.TestCase):
     def test_sub_tensor(self):
         shape = (3, 5)
         with nn.variable_scope(self.id().replace('.', '/')):
-            tensor1 = fixture.create_tensor(shape)
+            tensor1 = fixture.create_tensor(shape, dtype='int32')
             tensor2 = tensor1 * tensor1
 
         session = nn.Session()
@@ -179,7 +179,7 @@ class TestTensorOps(unittest.TestCase):
         shape = (3, 5)
         with nn.variable_scope(self.id().replace('.', '/')):
             input_ = nn.Input(shape=shape, dtype='int32')
-            tensor1 = fixture.create_tensor(shape)
+            tensor1 = fixture.create_tensor(shape, dtype='int32')
             tensor2 = tensor1 - input_
             tensor3 = input_ - tensor1
 
@@ -195,8 +195,8 @@ class TestTensorOps(unittest.TestCase):
     def test_sub_variable(self):
         shape = (3, 5)
         with nn.variable_scope(self.id().replace('.', '/')):
-            tensor1 = fixture.create_tensor(shape)
-            variable = fixture.create_variable(shape)
+            tensor1 = fixture.create_tensor(shape, dtype='int32')
+            variable = fixture.create_variable(shape, dtype='int32')
             tensor2 = tensor1 - variable
             tensor3 = variable - tensor1
